Cell type provides interior immutability
Because of the Rust constraints, a Type can only have either aliasing or
exlusive access and you can't  do them both. But in some cases, you need
a type where you want to be able to have multiple mutating access.
Example could be a datastructure like LinkedList where multiple items
have pointers to the same single object. Without a Cell, you can't
really do that as multiple owners can't have a mutating access.

**How is shared a mutating access even allowed?**

By using `UnsafeCell`, which allows to get a mutable reference on a
object but is considered unsafe by using the `get()` function.

**What the downsides of this in Multithreaded?**

Obviously If Cell type is used in multi-threaded programs, the threads
get raced and some of the modifications will be lost as there is no
locking mechanism.

This is prevented automatically as UnsafeCell does not implement Sync
Trait, causing it not to be able to be used btw threads. But in the
example, we implemented the Sync primitive unsafely and made it
possible, and if you run the test, the tests fail because of the race
updates.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>