RefCell is cell with Dynamic borrow checker rules
It uses rusts lifetimes to implement dynamic borrowing, through which
one can claim temporary mutable references to the inner value. These are
tracked at runtime, unlike Rust which does statically.

With RefCell we are essentially dynamically implementing the borrow
checking rules at runtime, useful in building graphs, etc where you
would want to track references to not do any racy behaviour.

RefCell allows the above by tracking all the shared and exclusive
references through using a refstate variable, which has a need of being
mutated with a shared reference, so we use Cell there which allows that.

We use two types i.e Ref and RefMut to make this possible, as we need a
way to track when we give back references of this i.e through borrow() and
borrow_mut(). We also implement DeRef for the above functions to make
accesing of the inner types direct. As deref is like the impl of .
operator.

signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>